feat(app): Further Capacitor work
