+++ app/templates/capacitor/www/index.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Quasar</title>
+
+    <meta charset="utf-8">
+    <meta name="description" content="Quasar Capacitor App">
+    <meta name="format-detection" content="telephone=no">
+    <meta name="msapplication-tap-highlight" content="no">
+    <meta name="viewport" content="user-scalable=no, initial-scale=1, maximum-scale=1, minimum-scale=1, width=device-width, viewport-fit=cover">
+  </head>
+
+  <body>
+    <div>
+      This file will be auto-generated. Do not edit.
+    </div>
+
+    <div>
+      Run "quasar dev" or "quasar build" with Capacitor mode.
+    </div>
+  </body>
+</html>
